+++ allocator/distances/__init__.py
@@ -1,7 +1,11 @@
 """Distance matrix calculations."""
 
 from .euclidean import euclidean_distance_matrix, latlon2xy, pairwise_distances, xy2latlog
-from .external_apis import google_distance_matrix, osrm_distance_matrix
+from .external_apis import (
+    google_distance_matrix,
+    google_routes_distance_matrix,
+    osrm_distance_matrix,
+)
 from .factory import get_distance_matrix
 from .haversine import haversine_distance_matrix
 
@@ -9,6 +13,7 @@ __all__ = [
     "euclidean_distance_matrix",
     "get_distance_matrix",
     "google_distance_matrix",
+    "google_routes_distance_matrix",
     "haversine_distance_matrix",
     "latlon2xy",
     "osrm_distance_matrix",